feat: classify errors based on log field
Resolves #37 by introducing a new pattern‑based `ErrorProcessor` that matches error entries in the
log payload and annotates events with the most recent error category.

- Renamed the former `ErrorProcessor` to `BlockReasonProcessor`.
- Updated the index template to include a new `block_reason` field and map block reasons there
  instead of `error_type`.
- Implemented a new `ErrorProcessor` which scans the log string against known error patterns and
  selects the error match appearing furthest in the text, ensuring the latest error is captured.
