+++ src/main/kotlin/ch/srgssr/pillarbox/monitoring/event/model/EventRequestDataConverter.kt
@@ -13,14 +13,17 @@ import com.fasterxml.jackson.databind.util.StdConverter
  * @see [DataProcessor]
  */
 internal class EventRequestDataConverter : StdConverter<EventRequest, EventRequest>() {
-  private val processors = listOf(UserAgentProcessor(), ErrorProcessor())
+  private val processors = listOf(UserAgentProcessor(), ContentRestrictionProcessor(), ErrorProcessor())
 
   @Suppress("UNCHECKED_CAST")
   override fun convert(value: EventRequest): EventRequest {
     (value.data as? MutableMap<String, Any?>)?.let { data ->
       processors
-        .filter { it.shouldProcess(value.eventName) }
-        .forEach { processor -> value.data = processor.process(data) }
+        .forEach { processor ->
+          if (processor.shouldProcess(value.eventName, data)) {
+            value.data = processor.process(data)
+          }
+        }
     }
 
     return value
